Fullpagina (#100)
* esquema

* modelo

* mapa funcional

* completo

+++ api2/src/models/safePlace.js
@@ -34,11 +34,11 @@ module.exports = function (sequelize) {
             allowNull:false,
         },
         lat: {
-            type: DataTypes.DOUBLE,
+            type: DataTypes.FLOAT,
             allowNull: true,
         },
         lng: {
-            type: DataTypes.DOUBLE,
+            type: DataTypes.FLOAT,
             allowNull: true,
         },
         email:{
